Change slate from package to module
